Cliffs, Beaches, etc.  Placement code for entities. -Cliffs and beaches are now implemented. -You can now place a test player using right click. -The placement code is set up to align with the tiles

+++ Vector2.cs
@@ -16,7 +16,7 @@ namespace EngineeringCorpsCS
             this.y = y;
         }
 
-        public Vector2(int[] xy)
+        public Vector2(float[] xy)
         {
             this.x = xy[0];
             this.y = xy[1];
